[Refactor] Feign 연동 버그 수정 및 리팩토링

+++ client/delivery/src/main/java/com/logistic/client/delivery/application/dto/HubRouteResponse.java
@@ -10,10 +10,10 @@ import java.util.UUID;
 @NoArgsConstructor
 @AllArgsConstructor
 public class HubRouteResponse {
-    private int sequence;
+    private Integer sequence;
     private UUID departureHubId;
     private UUID destinationHubId;
     private UUID deliveryManagerId;
-    private Integer distance;
-    private Integer time;
+    private double distance;
+    private double time;
 }
